requirements.txt
apex==0.9.10dev
numpy==1.21.2
opencv_python==4.5.4.58
Pillow==9.2.0
PyYAML==6.0
scikit_learn==1.1.2
spams==2.6.5.3
wandb==0.12.9
torch==1.10.0
torchvision==0.11.1
timm==0.4.12
einops==0.3.2
pandas==1.3.4
